refactor(gtk): move utils stuff to main and lib respectively
