+++ crates/client/src/lib.rs
@@ -1,3 +1,8 @@
 pub mod domain;
 mod error;
 
+pub fn version() -> String {
+    format!("{} v{}", env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"))
+        .trim()
+        .to_string()
+}
